test(flyway): move flyway integration tests into dedicated package
